requirements.txt
fastapi~=0.115.13
python-dotenv~=1.1.1
uvicorn~=0.34.3
pymongo~=4.13.2
pydantic~=2.11.7
redis~=6.2.0
PyJWT~=2.10.1